+++ dask/array/_array_expr/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
 
+import dask.array._array_expr._backends
 from dask.array._array_expr import random
 from dask.array._array_expr._collection import Array, blockwise, elemwise
