+++ dask/array/tests/test_fft.py
@@ -26,10 +26,12 @@ all_nd_funcnames = [
     "irfftn",
 ]
 
-nparr = np.arange(100).reshape(10, 10)
-darr = da.from_array(nparr, chunks=(1, 10))
-darr2 = da.from_array(nparr, chunks=(10, 1))
-darr3 = da.from_array(nparr, chunks=(10, 10))
+if not da._array_expr_enabled():
+
+    nparr = np.arange(100).reshape(10, 10)
+    darr = da.from_array(nparr, chunks=(1, 10))
+    darr2 = da.from_array(nparr, chunks=(10, 1))
+    darr3 = da.from_array(nparr, chunks=(10, 10))
 
 
 @pytest.mark.parametrize("funcname", all_1d_funcnames)
