+++ dask/dataframe/dask_expr/__init__.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from dask.dataframe.dask_expr import datasets
 from dask.dataframe.dask_expr._collection import *
-from dask.dataframe.dask_expr._dispatch import get_collection_type
 from dask.dataframe.dask_expr._dummies import get_dummies
 from dask.dataframe.dask_expr._groupby import Aggregation
 from dask.dataframe.dask_expr.io._delayed import from_delayed
